add support for zepto and, in theory, ender
wrap jasmine-jquery definition to pass in one of:
  - jQuery
  - Zepto

as much as possible, use shared API

about 20 specs are wrapped in guards to only run in certain conditions.  for example, with the optional `data` and `selectors` packages added to zepto, many more specs pass.  in the end, there are 6 remaining specs that do not work with any form of zepto (that I've found).  those have to do with:

  1. the execution of <script> elements in ajax responses… the script is executed, and there is a test for that, but it runs too soon to have access to the preceding fixture html
  2. toHandle and toHandleWith do not work because zepto does not expose the bound event handlers in any way.

some small amount of work has also been done toward supporting ender (based on the "jeesh" build), but the specs were still %50 red… maybe another time.

finally, I've kept the contents of jasmine-jquery.js fully left-justified in this commit (even though they're all wrapped in another function).  this is to make the diff easier to deal with.  the next commit will be a whitespace only adjustment.

simply toggle the comment blocks in SpecRunner.html to switch between jQuery and Zepto.