Do not trying to clone a non-DOM object
Previously, elementToString() function would trying to clone any object
passed to it for displaying in the test result. However, this would fail
if the user is trying to test anything against a non-DOM object, such as
a function. This commit make sure that the element is clonable before
attempting to clone it.