Match if the surrounding whitespace in the input matches the element text
Currently the text in the element is trimmed before comparing. This is helpful
in some cases, but we're unable to match intentional surrounding whitespace.

For instance, given an element like this:
<div> foo</div>
The following matcher will always fail:
expect($div).toHaveText(' foo');

This fixes that case so that the matcher passes if the surrounding whitespace
in the input matches the surrounding whitespace in the element text exactly.